Move classes to dedicated package
