+++ vector/src/main/java/im/vector/activity/VectorHomeActivity.java
@@ -119,9 +119,9 @@ import butterknife.BindView;
 import butterknife.OnClick;
 import fr.gouv.tchap.activity.TchapPublicRoomSelectionActivity;
 import fr.gouv.tchap.activity.TchapRoomCreationActivity;
-import fr.gouv.tchap.activity.TchapVersionCheckActivity;
-import fr.gouv.tchap.config.VersionCheckResult;
-import fr.gouv.tchap.config.VersionChecker;
+import fr.gouv.tchap.version.TchapVersionCheckActivity;
+import fr.gouv.tchap.version.VersionCheckResult;
+import fr.gouv.tchap.version.VersionChecker;
 import fr.gouv.tchap.fragments.TchapContactFragment;
 import fr.gouv.tchap.fragments.TchapRoomsFragment;
 import fr.gouv.tchap.util.DinsicUtils;
